requirements.txt
google-cloud-bigquery
requests